+++ src/app/globals.css
@@ -59,6 +59,13 @@ h1 {
   height: 100px;
 }
 
-.tarjetas {
-  background-color: #3E5C76;
+.tarjeta {
+  padding: 10px;
+  border: #1D2D44 1px solid;
+  margin: 5px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  border-radius: 10px;
+  background: #e0e0e0;
 }
